03_PROJECTS/youtube_automation/requirements.txt
# ============================================================
# YouTube自動化パイプライン — Python 依存パッケージ
# ============================================================

# Core
groq==0.4.2
pydantic==2.5.0
python-dotenv==1.0.0
pyyaml==6.0
requests==2.31.0

# Image & Video Processing
pillow==10.1.0
moviepy==1.0.3

# Audio
pydub==0.25.1
gTTS==2.5.0

# Google APIs
google-auth-oauthlib==1.2.0
google-api-python-client==2.107.0
google-auth-httplib2==0.2.0

# Data & API Clients
pexels-api==1.0.2
aiohttp==3.9.1

# AWS (Optional)
boto3==1.34.0
botocore==1.34.23

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Logging & Monitoring
python-json-logger==2.0.7

# Development (Optional)
pytest==7.4.3
black==23.12.0
flake8==6.1.0
